market/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
idna==3.4
importlib-resources==6.0.1
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
zipp==3.2
